Add convenience functions to generate source files and simple binaries.
Use cmake_unit_create_source_file_before_build to generate a source
file during the configure phase. A NAME for the source file can be
passed in as well as a list of INCLUDES, DEFINES and FUNCTIONS. Custom
contents can be added using PREPEND_CONTENTS.

Use cmake_unit_generate_source_file_during_build to generate a source
file in the build phase. This function takes similar arguments to
cmake_unit_create_source_file_before_build.

Use cmake_unit_create_simple_executable to create a simple executable
with the name NAME that just returns 0.

Use cmake_unit_create_simple_library with the name NAME and some
specified FUNCTIONS to create a simple library that has those
functions. The functions will be exported automatically.

If you need to get the full path to a compiled library or executable
in the verify phase, then use export (TARGETS ${TARGET}) in the
configure phase and then use
cmake_unit_get_target_location_from_exports specifying the location
of the EXPORTS file in the build-directory and the name of the target
to look up.